Create Periodic_3_triangulation_traits_3.
Add class Periodic_3_triangulation_traits_3.
Add class Periodic_3_triangulation_filtered__traits_3.
Put Traits_with_offsets_adaptor in a separate header.
Put Periodic_3_construct_point_3 in a separate header.
Update tests.
