+++ Periodic_3_triangulation_3/include/CGAL/Periodic_3_Delaunay_triangulation_filtered_traits_3.h
@@ -21,8 +21,8 @@
 //                 Manuel Caroli <Manuel.Caroli@sophia.inria.fr>
 
 
-#ifndef CGAL_PERIODIC_3_TRIANGULATION_FILTERED_TRAITS_3_H
-#define CGAL_PERIODIC_3_TRIANGULATION_FILTERED_TRAITS_3_H
+#ifndef CGAL_PERIODIC_3_DELAUNAY_TRIANGULATION_FILTERED_TRAITS_3_H
+#define CGAL_PERIODIC_3_DELAUNAY_TRIANGULATION_FILTERED_TRAITS_3_H
 
 #include <string>
 #include <CGAL/basic.h>
@@ -31,74 +31,13 @@
 #include <CGAL/Uncertain.h>
 #include <CGAL/Profile_counter.h>
 #include <CGAL/Periodic_3_Delaunay_triangulation_traits_3.h>
+#include <CGAL/Periodic_3_triangulation_filtered_traits_3.h>
 
 namespace CGAL {
 
-// This template class is a wrapper that implements the filtering for any
-// predicate (dynamic filters with IA).
-
-// TODO :
-// - each predicate in the default kernel should define a tag that says if it
-//   wants to be filtered or not (=> all homogeneous predicate define this
-//   tag).  We could even test-suite that automatically.  It makes a strong
-//   new requirement on the kernel though...
-//   Could be done with a traits mechanism ?
-//   A default template could use the current IA, but other tags or whatever
-//   could specify no filtering at all, or static filtering...
-// - same thing for constructions => virtual operator() ?
-// - similarly, constructions should have a tag saying if they can throw or
-//   not, or we let all this up to the compiler optimizer to figure out ?
-// - Some caching could be done at the Point_2 level.
-
-template <class EP, class AP, class C2E, class C2A, bool Protection = true>
-class Filtered_periodic_predicate
-  : public Filtered_predicate<EP, AP, C2E, C2A, Protection>
-{
-  typedef Filtered_predicate<EP, AP, C2E, C2A, Protection> Base;
-public:
-  Filtered_periodic_predicate() : Base() {}
-
-  // These constructors are used for constructive predicates.
-  // You should try to avoid constructive predicates, as they will construct
-  // the exact values systematically (in the ctor), rather than lazily.
-  template <class OE, class OA> 
-  Filtered_periodic_predicate(const OE * oe, const OA * oa) 
-    : Base( EP(oe), AP(oa) )
-  {}
-};
-
-// The Offset_converter is parametrized by a usual kernel converter,
-// and adds the conversions for Offsets.
-template < typename Converter >
-struct Offset_converter_3
-  : public Converter
-{
-  typedef typename Converter::Source_kernel Source_kernel;
-  typedef typename Converter::Target_kernel Target_kernel;
-
-  typedef typename Periodic_3_Delaunay_triangulation_traits_base_3<Source_kernel>
-                   ::Offset  Source_off;
-  typedef typename Periodic_3_Delaunay_triangulation_traits_base_3<Source_kernel>
-                   ::Point_3  Source_pt;
-
-  typedef typename Periodic_3_Delaunay_triangulation_traits_base_3<Target_kernel>
-                   ::Offset  Target_off;
-  typedef typename Periodic_3_Delaunay_triangulation_traits_base_3<Target_kernel>
-                   ::Point_3  Target_pt;
-
-
-  using Converter::operator();
-
-  Target_off
-  operator()(const Source_off &off) const
-  {
-    return off;
-  }
-};
-
 // The argument is supposed to be a Filtered_kernel like kernel.
 template < typename K, typename Off >
-class Periodic_3_triangulation_filtered_traits_base_3
+class Periodic_3_Delaunay_triangulation_filtered_traits_base_3
   : public Periodic_3_Delaunay_triangulation_traits_base_3<K, Off>
 {
   typedef Periodic_3_Delaunay_triangulation_traits_base_3<K, Off> Base;
@@ -214,9 +153,9 @@ namespace CGAL {
 template < typename K, typename Off = typename CGAL::Periodic_3_offset_3 >
 class Periodic_3_Delaunay_triangulation_filtered_traits_3
   : public Periodic_3_triangulation_statically_filtered_traits_3<
-  Periodic_3_triangulation_filtered_traits_base_3<K, Off> > {
+  Periodic_3_Delaunay_triangulation_filtered_traits_base_3<K, Off> > {
 };
 
 } //namespace CGAL
 
-#endif // CGAL_PERIODIC_3_TRIANGULATION_FILTERED_TRAITS_3_H
+#endif // CGAL_PERIODIC_3_DELAUNAY_TRIANGULATION_FILTERED_TRAITS_3_H
