+++ Periodic_3_triangulation_3/test/Periodic_3_triangulation_3/test_periodic_3_triangulation_3.cpp
@@ -24,25 +24,30 @@
 
 #include <CGAL/Periodic_3_Delaunay_triangulation_3.h>
 #include <CGAL/Periodic_3_Delaunay_triangulation_traits_3.h>
+#include <CGAL/Periodic_3_triangulation_3.h>
+#include <CGAL/Periodic_3_triangulation_traits_3.h>
 
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 typedef CGAL::Exact_predicates_inexact_constructions_kernel K1;
-typedef CGAL::Periodic_3_Delaunay_triangulation_traits_3<K1> PTT1;
+typedef CGAL::Periodic_3_triangulation_traits_3<K1> PTT1;
+typedef CGAL::Periodic_3_Delaunay_triangulation_traits_3<K1> PDTT1;
 // Explicit instantiation of the whole class:
-template class CGAL::Periodic_3_triangulation_3<PTT1>;
+template class CGAL::Periodic_3_triangulation_3<PDTT1>;
 
 #include <CGAL/Exact_predicates_exact_constructions_kernel.h>
 typedef CGAL::Exact_predicates_exact_constructions_kernel K2;
-typedef CGAL::Periodic_3_Delaunay_triangulation_traits_3<K2> PTT2;
+typedef CGAL::Periodic_3_triangulation_traits_3<K2> PTT2;
+typedef CGAL::Periodic_3_Delaunay_triangulation_traits_3<K2> PDTT2;
 // Explicit instantiation of the whole class:
-template class CGAL::Periodic_3_triangulation_3<PTT2>;
+template class CGAL::Periodic_3_triangulation_3<PDTT2>;
 
 #include <CGAL/MP_Float.h>
 #include <CGAL/Simple_homogeneous.h>
 typedef CGAL::Simple_homogeneous<CGAL::MP_Float> K3;
-typedef CGAL::Periodic_3_Delaunay_triangulation_traits_3<K3> PTT3;
+typedef CGAL::Periodic_3_triangulation_traits_3<K3> PTT3;
+typedef CGAL::Periodic_3_Delaunay_triangulation_traits_3<K3> PDTT3;
 // Explicit instantiation of the whole class:
-template class CGAL::Periodic_3_triangulation_3<PTT3>;
+template class CGAL::Periodic_3_triangulation_3<PDTT3>;
 
 #include <CGAL/_test_cls_periodic_3_triangulation_3.h>
 
@@ -55,12 +60,12 @@ int main()
 {
   typedef CGAL::Periodic_3_triangulation_3<PTT1>            P3T3_1;
   _test_periodic_3_triangulation_3_constructors( P3T3_1() );
-  typedef CGAL::Periodic_3_Delaunay_triangulation_3<PTT1>   P3DT3_1;
+  typedef CGAL::Periodic_3_Delaunay_triangulation_3<PDTT1>   P3DT3_1;
   _test_cls_periodic_3_triangulation_3( P3DT3_1() );
 
   typedef CGAL::Periodic_3_triangulation_3<PTT2>            P3T3_2;
   _test_periodic_3_triangulation_3_constructors( P3T3_2() );
-  typedef CGAL::Periodic_3_Delaunay_triangulation_3<PTT2>   P3DT3_2;
+  typedef CGAL::Periodic_3_Delaunay_triangulation_3<PDTT2>   P3DT3_2;
   _test_cls_periodic_3_triangulation_3( P3DT3_2(), true );
 
   typedef CGAL::Periodic_3_triangulation_3<PTT3>            P3T3_3;
